requirement.txt
pyspark==3.2.1
kafka-python==2.0.2
tweepy==4.10.0